rework make_consistent
