+++ interfaces/acados_template/acados_template/acados_model.py
@@ -28,8 +28,13 @@
 # POSSIBILITY OF SUCH DAMAGE.;
 #
 
+from typing import Union
+
 from casadi import MX, SX
-from .utils import is_empty
+
+from .utils import is_empty, casadi_length, is_column
+from .acados_dims import AcadosOcpDims, AcadosSimDims
+
 
 class AcadosModel():
     """
@@ -204,7 +209,7 @@ class AcadosModel():
         Used if :py:attr:`acados_template.acados_ocp.AcadosOcpOptions.cost_type_e` is 'CONVEX_OVER_NONLINEAR'.
         """
 
-    def make_consistent(self) -> None:
+    def make_consistent(self, dims: Union[AcadosOcpDims, AcadosSimDims]) -> None:
         x = self.x
         z = self.z
         p = self.p
@@ -221,4 +226,29 @@ class AcadosModel():
 
         if is_empty(z):
             self.z = symbol('z', 0, 0)
+
+        # nx
+        if is_column(self.x):
+            dims.nx = casadi_length(self.x)
+        else:
+            raise Exception('model.x should be column vector!')
+
+        # nu
+        if is_empty(self.u):
+            dims.nu = 0
+        else:
+            dims.nu = casadi_length(self.u)
+
+        # nz
+        if is_empty(self.z):
+            dims.nz = 0
+        else:
+            dims.nz = casadi_length(self.z)
+
+        # np
+        if is_empty(self.p):
+            dims.np = 0
+        else:
+            dims.np = casadi_length(self.p)
+
         return
