+++ interfaces/acados_template/acados_template/acados_ocp_solver.py
@@ -53,7 +53,7 @@ from .acados_ocp import AcadosOcp
 from .acados_model import AcadosModel
 from .acados_ocp_cost import AcadosOcpCost
 from .acados_ocp_constraints import AcadosOcpConstraints
-from .acados_ocp_dims import AcadosOcpDims
+from .acados_dims import AcadosOcpDims
 from .acados_ocp_options import AcadosOcpOptions
 
 from .builders import CMakeBuilder
