+++ interfaces/acados_template/acados_template/acados_sim.py
@@ -29,72 +29,11 @@
 # POSSIBILITY OF SUCH DAMAGE.;
 #
 
-import numpy as np
 import os
+import numpy as np
 from .acados_model import AcadosModel
+from .acados_dims import AcadosSimDims
 from .utils import get_acados_path, get_shared_lib_ext, casadi_length, is_empty, is_column
-
-class AcadosSimDims:
-    """
-    Class containing the dimensions of the model to be simulated.
-    """
-    def __init__(self):
-        self.__nx = None
-        self.__nu = None
-        self.__nz = 0
-        self.__np = 0
-
-    @property
-    def nx(self):
-        """:math:`n_x` - number of states. Type: int > 0"""
-        return self.__nx
-
-    @property
-    def nz(self):
-        """:math:`n_z` - number of algebraic variables. Type: int >= 0"""
-        return self.__nz
-
-    @property
-    def nu(self):
-        """:math:`n_u` - number of inputs. Type: int >= 0"""
-        return self.__nu
-
-    @property
-    def np(self):
-        """:math:`n_p` - number of parameters. Type: int >= 0"""
-        return self.__np
-
-    @nx.setter
-    def nx(self, nx):
-        if isinstance(nx, int) and nx > 0:
-            self.__nx = nx
-        else:
-            raise Exception('Invalid nx value, expected positive integer.')
-
-    @nz.setter
-    def nz(self, nz):
-        if isinstance(nz, int) and nz > -1:
-            self.__nz = nz
-        else:
-            raise Exception('Invalid nz value, expected nonnegative integer.')
-
-    @nu.setter
-    def nu(self, nu):
-        if isinstance(nu, int) and nu > -1:
-            self.__nu = nu
-        else:
-            raise Exception('Invalid nu value, expected nonnegative integer.')
-
-    @np.setter
-    def np(self, np):
-        if isinstance(np, int) and np > -1:
-            self.__np = np
-        else:
-            raise Exception('Invalid np value, expected nonnegative integer.')
-
-    def set(self, attr, value):
-        setattr(self, attr, value)
-
 
 class AcadosSimOpts:
     """
@@ -357,31 +296,8 @@ class AcadosSim:
     def make_consistent(self):
         dims = self.dims
         model = self.model
-        model.make_consistent()
+        model.make_consistent(dims)
 
-        # nx
-        if is_column(model.x):
-            dims.nx = casadi_length(model.x)
-        else:
-            raise Exception('model.x should be column vector!')
-
-        # nu
-        if is_empty(model.u):
-            dims.nu = 0
-        else:
-            dims.nu = casadi_length(model.u)
-
-        # nz
-        if is_empty(model.z):
-            dims.nz = 0
-        else:
-            dims.nz = casadi_length(model.z)
-
-        # np
-        if is_empty(model.p):
-            dims.np = 0
-        else:
-            dims.np = casadi_length(model.p)
         if self.parameter_values.shape[0] != dims.np:
             raise Exception('inconsistent dimension np, regarding model.p and parameter_values.' + \
                 f'\nGot np = {dims.np}, acados_sim.parameter_values.shape = {self.parameter_values.shape[0]}\n')
